Separate game_logic from game_state (data and immutable queries).
