+++ royals_core/src/lib.rs
@@ -5,6 +5,7 @@ use player::{Player, PlayerId};
 pub mod card;
 pub mod event;
 mod game_state;
+mod game_logic;
 pub mod play;
 pub mod player;
 mod random_playing_computer;
